+++ src/planetarypy/instruments/mro/ctx/__init__.py
@@ -6,16 +6,18 @@ Example usage:
     from planetarypy.instruments.mro.ctx import CTXCollection, EDR, Calib
 """
 
+from planetarypy.instruments.mro.ctx.ctx_edr import (
+    EDR,
+    get_edr_index,
+    product_id_from_serial_number,
+)
 from planetarypy.instruments.mro.ctx.ctx_calib import (
     Calib,
     CTXCollection,
-    EDR,
     calibrate_pid,
     do_footprintinit,
     download_pid,
-    get_edr_index,
     process_parallel,
-    product_id_from_serial_number,
 )
 
 __all__ = [